Add selection of actual setpoint in temperature setting container

+++ main/GuiApp/InfoScreen.cpp
@@ -70,8 +70,8 @@ void InfoScreen::init(){
 }
 
 void InfoScreen::run(){
-	uint16_t actual = GuiApp::getActual();
-	if (actual == (uint16_t)GuiApp::InvalidValue) {
+	int16_t actual = GuiApp::getActual();
+	if (actual == GuiApp::InvalidValue) {
 		lv_label_set_text_fmt(mActualTempLabel, "--°C");
 	} else {
 		lv_label_set_text_fmt(mActualTempLabel, "%d.%d°C", actual/10, actual%10);
